+++ usi/api/systemc.h
@@ -38,7 +38,7 @@ double delta_count();
 
 int set_verbosity(int verbosity);
 
-long get_curr_process_handle();
+size_t get_curr_process_handle();
 bool is_running();
 std::vector<sc_core::sc_object *> get_top_level_objects();
